Refactor: CurrenciesDataSlice: rename currencySlice to currenciesDataSlice
