+++ src/components/Currency/CurrenciesList.tsx
@@ -5,7 +5,7 @@ import { apiState } from 'types/global';
 
 import CurrencyItem from 'components/Currency/CurrencyItem';
 import { getBaseCurrency, getFilteredCurrencies, getLoadingStatus } from 'components/Currency/redux/currencySelector';
-import { currenciesActions, fetchCurrencies } from 'components/Currency/redux/currencySlice';
+import { currenciesActions, fetchCurrencies } from 'components/Currency/redux/currenciesDataSlice';
 
 import { searchSelector } from 'components/SearchBar/searchBarSelectors';
