+++ src/components/Currency/redux/currenciesDataSlice.ts
